requirements.txt
addict==2.4.0
astunparse==1.6.3
cython==3.0.10
diffusers==0.29.2
ffmpy==0.4.0
Flask==3.0.3
gdown==5.2.0
huggingface_hub>=0.25.1
hydra-core==1.3.2
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
matplotlib==3.9.1
mediapipe==0.10.15
mmcv==2.2.0
mmengine==0.10.5
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.20.0
openai==1.54.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
progressbar33==2.4
psutil==6.0.0
pybullet==3.2.6
pycocotools==2.0.8
Pygments==2.15.1
PyYAML==6.0.1
Requests==2.32.3
scipy==1.14.1
setuptools==60.2.0
Shapely==2.0.6
scikit-image==0.24.0
supervision==0.24.0
termcolor==2.5.0
timm==1.0.11
tqdm==4.66.6
transformers==4.43.3
yapf==0.40.2
httpx==0.27.2